NeuroSign-main/requirements.txt
Flask==3.1.0
SpeechRecognition==3.14.1
pyaudio==0.2.14
opencv-python
mediapipe
numpy

Flask
Flask-Cors
pymongo
python-dotenv
google-generativeai
random